Moved Minecart related classes to new package and renamed VehicleDropItemEventListener to VehicleDestroyEventListener
